pyproject.toml
[tool.poetry]
name = "verivynx"
version = "0.1.0"
description = "Verification Bot for Vynx"
authors = ["timelessnesses <mooping3roblox@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.7.3"
aiohttp = {extras = ["speedups"], version = "<3.8.0"}
python-dotenv = "^0.19.2"


[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^21.12b0"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
